refactor(libkrun-sys): split the safe wrapper lib.rs into submodules
lib.rs was a ~2655-line crate root mixing install probing, KrunContext,
the boot path, the bridge-inserting boot path, and SupervisorConfig.
Move the wrapper body into error.rs, context.rs, start.rs, bridge.rs,
and supervisor.rs; lib.rs is now module decls plus named re-exports so
libkrun_sys::<Name> resolves identically for every caller. Pure
structural move, no behavior change.
